extraccion-documentos-generales/requirements.txt
azure-ai-formrecognizer==3.3.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-identity
azure-keyvault-secrets
azure-common==1.1.28
azure-core
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2024.07.04
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.1
Flask-Jsonpify==1.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.7
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.4
lxml==4.9.3
MarkupSafe==2.1.3
msrest==0.7.1
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
passlib==1.7.4
pyodbc==5.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.32.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.2.2
user-agents==2.2.0
Werkzeug==3.0.3
zipp==3.19.1
WTForms==3.1.2
Flask-WTF==1.2.1